+++ docs/release_notes.rst
@@ -6,6 +6,7 @@ Release Notes
 -----
 
 * Supported native modules with rebuilt scripts
+* Fixed ``NodeRuntime.await()`` to emit tasks
 
 0.8.6
 -----